requirements/base.txt
coloredlogs==15.0.1
Jinja2==3.1.3
langchain==0.1.13
langchain_community==0.0.29
langchain_core==0.1.33
langchain_openai==0.1.0
mistletoe==1.3.0
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-chromadb==0.15.11
opentelemetry-instrumentation-langchain==0.15.11
opentelemetry-instrumentation-transformers==0.15.11
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-semantic-conventions-ai==0.1.1
opentelemetry-util-http==0.45b0
pandas==2.2.1
pillow==10.2.0
playwright==1.42.0
pydantic==2.6.4
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
tokenizers==0.15.2
torch==2.2.2
traceloop-sdk==0.15.11
transformers==4.39.3
websocket-client==1.7.0
